+++ js-api-spec/importer.test.ts
@@ -7,154 +7,6 @@ import {compile, compileString, compileStringAsync, Importer} from 'sass';
 
 import {sandbox} from './utils_node';
 
-it('uses an importer to resolve an @import', () => {
-  const result = compileString('@import "orange";', {
-    importers: [
-      {
-        canonicalize: url => new URL(`u:${url}`),
-        load(url) {
-          const color = url.pathname;
-          return {contents: `.${color} {color: ${color}}`, syntax: 'scss'};
-        },
-      },
-    ],
-  });
-
-  expect(result.css).toBe('.orange {\n  color: orange;\n}');
-});
-
-it('passes the canonicalized URL to the importer', () => {
-  const result = compileString('@import "orange";', {
-    importers: [
-      {
-        canonicalize: () => new URL('u:blue'),
-        load(url) {
-          const color = url.pathname;
-          return {contents: `.${color} {color: ${color}}`, syntax: 'scss'};
-        },
-      },
-    ],
-  });
-
-  expect(result.css).toBe('.blue {\n  color: blue;\n}');
-});
-
-it('only invokes the importer once for a given canonicalization', () => {
-  const result = compileString(
-    `
-      @import "orange";
-      @import "orange";
-    `,
-    {
-      importers: [
-        {
-          canonicalize: () => new URL('u:blue'),
-          load(url) {
-            const color = url.pathname;
-            return {contents: `.${color} {color: ${color}}`, syntax: 'scss'};
-          },
-        },
-      ],
-    }
-  );
-
-  expect(result.css).toBe(`.blue {
-  color: blue;
-}
-
-.blue {
-  color: blue;
-}`);
-});
-
-describe('the imported URL', () => {
-  // Regression test for sass/dart-sass#1137.
-  it("isn't changed if it's root-relative", () => {
-    const result = compileString('@import "/orange";', {
-      importers: [
-        {
-          canonicalize(url) {
-            expect(url).toEqual('/orange');
-            return new URL(`u:${url}`);
-          },
-          load: () => ({contents: 'a {b: c}', syntax: 'scss'}),
-        },
-      ],
-    });
-
-    expect(result.css).toBe('a {\n  b: c;\n}');
-  });
-
-  it("is converted to a file: URL if it's an absolute Windows path", () => {
-    const result = compileString('@import "C:/orange";', {
-      importers: [
-        {
-          canonicalize(url) {
-            expect(url).toEqual('file:///C:/orange');
-            return new URL(`u:${url}`);
-          },
-          load: () => ({contents: 'a {b: c}', syntax: 'scss'}),
-        },
-      ],
-    });
-
-    expect(result.css).toBe('a {\n  b: c;\n}');
-  });
-});
-
-it("uses an importer's source map URL", () => {
-  const result = compileString('@import "orange";', {
-    importers: [
-      {
-        canonicalize: url => new URL(`u:${url}`),
-        load(url) {
-          const color = url.pathname;
-          return {
-            contents: `.${color} {color: ${color}}`,
-            syntax: 'scss',
-            sourceMapUrl: new URL('u:blue'),
-          };
-        },
-      },
-    ],
-    sourceMap: true,
-  });
-
-  expect(result.sourceMap!.sources).toInclude('u:blue');
-});
-
-it('wraps an error in canonicalize()', () => {
-  expect(() => {
-    compileString('@import "orange";', {
-      importers: [
-        {
-          canonicalize() {
-            throw 'this import is bad actually';
-          },
-          load() {
-            fail('load() should not be called');
-          },
-        },
-      ],
-    });
-  }).toThrowSassException({line: 0});
-});
-
-it('wraps an error in load()', () => {
-  expect(() => {
-    compileString('@import "orange";', {
-      importers: [
-        {
-          canonicalize: url => new URL(`u:${url}`),
-          load() {
-            throw 'this import is bad actually';
-          },
-        },
-      ],
-    });
-  }).toThrowSassException({line: 0});
-});
-
 it('avoids importer when canonicalize() returns null', () =>
   sandbox(dir => {
     dir.write({'dir/_other.scss': 'a {from: dir}'});
@@ -212,27 +64,6 @@ it('prefers a relative file load to an importer', () =>
     expect(result.css).toBe('a {\n  from: relative;\n}');
   }));
 
-it('prefers a relative importer load to an importer', () => {
-  const result = compileString('@import "other";', {
-    importers: [
-      {
-        canonicalize() {
-          throw 'canonicalize() should not be called';
-        },
-        load() {
-          throw 'load() should not be called';
-        },
-      },
-    ],
-    url: new URL('o:style.scss'),
-    importer: {
-      canonicalize: url => new URL(url),
-      load: () => ({contents: 'a {from: relative}', syntax: 'scss'}),
-    },
-  });
-  expect(result.css).toBe('a {\n  from: relative;\n}');
-});
-
 it('prefers an importer to a load path', () =>
   sandbox(dir => {
     dir.write({
@@ -251,167 +82,6 @@ it('prefers an importer to a load path', () =>
     });
     expect(result.css).toBe('a {\n  from: importer;\n}');
   }));
-
-describe('with syntax', () => {
-  it('scss, parses it as SCSS', () => {
-    const result = compileString('@import "other";', {
-      importers: [
-        {
-          canonicalize: () => new URL('u:other'),
-          load: () => ({contents: '$a: value; b {c: $a}', syntax: 'scss'}),
-        },
-      ],
-    });
-
-    expect(result.css).toBe('b {\n  c: value;\n}');
-  });
-
-  it('indented, parses it as the indented syntax', () => {
-    const result = compileString('@import "other";', {
-      importers: [
-        {
-          canonicalize: () => new URL('u:other'),
-          load: () => ({
-            contents: '$a: value\nb\n  c: $a',
-            syntax: 'indented',
-          }),
-        },
-      ],
-    });
-
-    expect(result.css).toBe('b {\n  c: value;\n}');
-  });
-
-  it('css, allows plain CSS', () => {
-    const result = compileString('@import "other";', {
-      importers: [
-        {
-          canonicalize: () => new URL('u:other'),
-          load: () => ({contents: 'a {b: c}', syntax: 'css'}),
-        },
-      ],
-    });
-
-    expect(result.css).toBe('a {\n  b: c;\n}');
-  });
-
-  it('css, rejects SCSS', () => {
-    expect(() => {
-      compileString('@import "other";', {
-        importers: [
-          {
-            canonicalize: () => new URL('u:other'),
-            load: () => ({contents: '$a: value\nb\n  c: $a', syntax: 'css'}),
-          },
-        ],
-      });
-    }).toThrowSassException({
-      line: 0,
-      url: new URL('u:other'),
-    });
-  });
-});
-
-describe('async', () => {
-  it('resolves an @import', async () => {
-    const result = await compileStringAsync('@import "orange";', {
-      importers: [
-        {
-          canonicalize: url => Promise.resolve(new URL(`u:${url}`)),
-          load(url) {
-            const color = url.pathname;
-            return Promise.resolve({
-              contents: `.${color} {color: ${color}}`,
-              syntax: 'scss',
-            });
-          },
-        },
-      ],
-    });
-
-    expect(result.css).toBe('.orange {\n  color: orange;\n}');
-  });
-
-  it('wraps an asynchronous error in canonicalize', async () => {
-    await expect(() =>
-      compileStringAsync('@import "orange";', {
-        importers: [
-          {
-            canonicalize: () => Promise.reject('this import is bad actually'),
-            load() {
-              fail('load() should not be called');
-            },
-          },
-        ],
-      })
-    ).toThrowSassException({line: 0});
-  });
-
-  it('wraps a synchronous error in canonicalize', async () => {
-    await expect(() =>
-      compileStringAsync('@import "orange";', {
-        importers: [
-          {
-            canonicalize() {
-              throw 'this import is bad actually';
-            },
-            load() {
-              fail('load() should not be called');
-            },
-          },
-        ],
-      })
-    ).toThrowSassException({line: 0});
-  });
-
-  it('wraps an asynchronous error in load', async () => {
-    await expect(() =>
-      compileStringAsync('@import "orange";', {
-        importers: [
-          {
-            canonicalize: url => new URL(`u:${url}`),
-            load: () => Promise.reject('this import is bad actually'),
-          },
-        ],
-      })
-    ).toThrowSassException({line: 0});
-  });
-
-  it('wraps a synchronous error in load', async () => {
-    await expect(() =>
-      compileStringAsync('@import "orange";', {
-        importers: [
-          {
-            canonicalize: url => new URL(`u:${url}`),
-            load() {
-              throw 'this import is bad actually';
-            },
-          },
-        ],
-      })
-    ).toThrowSassException({line: 0});
-  });
-});
-
-describe('fromImport is', () => {
-  it('true from an @import', () => {
-    compileString('@import "foo"', {importers: [expectFromImport(true)]});
-  });
-
-  it('false from a @use', () => {
-    compileString('@use "foo"', {importers: [expectFromImport(false)]});
-  });
-
-  it('false from a @forward', () => {
-    compileString('@forward "foo"', {importers: [expectFromImport(false)]});
-  });
-
-  it('false from meta.load-css', () => {
-    compileString('@use "sass:meta"; @include meta.load-css("")', {
-      importers: [expectFromImport(false)],
-    });
-  });
-});
 
 describe('FileImporter', () => {
   it('loads a fully canonicalized URL', () =>
@@ -658,87 +328,3 @@ it(
       expect(callback).not.toThrowSassException();
     })
 );
-
-describe('when importer does not return string contents', () => {
-  it('throws an error in sync mode', () => {
-    expect(() => {
-      compileString('@import "other";', {
-        importers: [
-          {
-            canonicalize: url => new URL(`u:${url}`),
-            load() {
-              return {
-                // Need to force an invalid type to test bad-type handling.
-                // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                contents: Buffer.from('not a string') as any,
-                syntax: 'scss',
-              };
-            },
-          },
-        ],
-      });
-    }).toThrowSassException({
-      line: 0,
-      includes: 'Invalid argument (contents): must be a string but was: Buffer',
-    });
-  });
-
-  it('throws an error in async mode', async () => {
-    await expect(async () => {
-      await compileStringAsync('@import "other";', {
-        importers: [
-          {
-            canonicalize: url => new URL(`u:${url}`),
-            load() {
-              return {
-                // Need to force an invalid type to test bad-type handling.
-                // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                contents: Buffer.from('not a string') as any,
-                syntax: 'scss',
-              };
-            },
-          },
-        ],
-      });
-    }).toThrowSassException({
-      line: 0,
-      includes: 'Invalid argument (contents): must be a string but was: Buffer',
-    });
-  });
-});
-
-it('throws an ArgumentError when the result sourceMapUrl is missing a scheme', () => {
-  expect(() => {
-    compileString('@import "other";', {
-      importers: [
-        {
-          canonicalize: url => new URL(`u:${url}`),
-          load() {
-            return {
-              contents: '',
-              syntax: 'scss',
-              sourceMapUrl: {} as URL,
-            };
-          },
-        },
-      ],
-    });
-  }).toThrowSassException({
-    line: 0,
-    includes: 'Invalid argument (sourceMapUrl): must be absolute',
-  });
-});
-
-/**
- * Returns an importer that asserts that `fromImport` is `expected`, and
- * otherwise imports exclusively empty stylesheets.
- */
-function expectFromImport(expected: boolean): Importer<'sync'> {
-  return {
-    canonicalize(url, {fromImport}) {
-      expect(fromImport).toBe(expected);
-      return new URL(`u:${url}`);
-    },
-    load: () => ({contents: '', syntax: 'scss'}),
-  };
-}
